+++ src/redi/cli/init_command.py
@@ -5,7 +5,8 @@ import requests
 from prompt_toolkit import prompt
 from prompt_toolkit.validation import Validator
 
-from redi.cli.prompt_util import UrlValidator, inline_choice
+from redi.cli.prompt_util import inline_choice
+from redi.cli.validator import UrlValidator
 from redi.config import CONFIG_PATH, create_profile
 from redi.i18n import messages
