+++ src/redi/cli/time_entry_command.py
@@ -6,13 +6,13 @@ from prompt_toolkit.validation import Validator
 
 from redi.cli.alias import resolve_alias
 from redi.cli.prompt_util import (
-    HourValidator,
     confirm_delete,
     digit_and_period_key_bindings,
     digit_only_key_bindings,
     inline_checkbox,
     inline_choice,
 )
+from redi.cli.validator import HourValidator
 from redi.config import default_project_id
 from redi.i18n import messages
 from redi.api.enumeration import fetch_time_entry_activities
